refactor(rag): extract temporal detection into rag/temporal.rs
Move DateRange, detect_temporal_regex, detect_temporal_llm, apply_date_filter
into temporal.rs. Drop now-unused chrono and TEMPORAL_DETECT_PROMPT imports
from mod.rs. Behavior unchanged.
